ref(tests): Rename `mocking` to `mock_endpoint_builder`
The new name more precisely defines the module's purpose.
